requirements.txt
asgiref==3.7.2
crispy-tailwind==1.0.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
